requirements.txt
python-dotenv==1.0.1
streamlit==1.44.1
requests==2.31.0
fastapi==0.110.0
uvicorn==0.30.1
google-generativeai==0.4.0
pydantic==2.7.1
loguru==0.7.2
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
plotly==5.23.0
seaborn==0.13.2
simpy==4.0.2
statsmodels==0.14.4
